+++ OUDS/Core/Components/Sources/Tags/Internal/TagLabelStyle+BackgroundModifier.swift
@@ -2,13 +2,13 @@
 // Software Name: OUDS iOS
 // SPDX-FileCopyrightText: Copyright (c) Orange SA
 // SPDX-License-Identifier: MIT
-// 
+//
 // This software is distributed under the MIT license,
 // the text of which is available at https://opensource.org/license/MIT/
 // or see the "LICENSE" file for more details.
-// 
+//
 // Authors: See CONTRIBUTORS.txt
-// Software description: A SwiftUI components library with code examples for Orange Unified Design System 
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
 //
 
 import OUDSTokensSemantic
@@ -16,10 +16,11 @@ import SwiftUI
 
 struct TagBackgroundModifier: ViewModifier {
 
-    // MARK: Stored Properties
+    // MARK: Stored properties
 
     let hierarchy: OUDSTag.Hierarchy
     let status: OUDSTag.Status
+
     @Environment(\.theme) private var theme
 
     // MARK: Body
@@ -33,47 +34,47 @@ struct TagBackgroundModifier: ViewModifier {
     private var background: MultipleColorSemanticTokens {
         switch hierarchy {
         case .emphasized:
-            return emphasizedBackground
+            emphasizedBackground
         case .muted:
-            return mutedBackground
+            mutedBackground
         }
     }
 
     private var emphasizedBackground: MultipleColorSemanticTokens {
         switch status {
         case .neutral:
-            return theme.colors.colorSurfaceStatusNeutralEmphasized
+            theme.colors.colorSurfaceStatusNeutralEmphasized
         case .accent:
-            return theme.colors.colorSurfaceStatusAccentEmphasized
+            theme.colors.colorSurfaceStatusAccentEmphasized
         case .positive:
-            return theme.colors.colorSurfaceStatusPositiveEmphasized
+            theme.colors.colorSurfaceStatusPositiveEmphasized
         case .warning:
-            return theme.colors.colorSurfaceStatusWarningEmphasized
+            theme.colors.colorSurfaceStatusWarningEmphasized
         case .negative:
-            return theme.colors.colorSurfaceStatusNegativeEmphasized
+            theme.colors.colorSurfaceStatusNegativeEmphasized
         case .info:
-            return theme.colors.colorSurfaceStatusInfoEmphasized
+            theme.colors.colorSurfaceStatusInfoEmphasized
         case .disabled:
-            return theme.colors.colorActionDisabled
+            theme.colors.colorActionDisabled
         }
     }
 
     private var mutedBackground: MultipleColorSemanticTokens {
         switch status {
         case .neutral:
-            return theme.colors.colorSurfaceStatusNeutralMuted
+            theme.colors.colorSurfaceStatusNeutralMuted
         case .accent:
-            return theme.colors.colorSurfaceStatusAccentMuted
+            theme.colors.colorSurfaceStatusAccentMuted
         case .positive:
-            return theme.colors.colorSurfaceStatusPositiveMuted
+            theme.colors.colorSurfaceStatusPositiveMuted
         case .warning:
-            return theme.colors.colorSurfaceStatusWarningMuted
+            theme.colors.colorSurfaceStatusWarningMuted
         case .negative:
-            return theme.colors.colorSurfaceStatusNegativeMuted
+            theme.colors.colorSurfaceStatusNegativeMuted
         case .info:
-            return theme.colors.colorSurfaceStatusInfoMuted
+            theme.colors.colorSurfaceStatusInfoMuted
         case .disabled:
-            return theme.colors.colorActionDisabled
+            theme.colors.colorActionDisabled
         }
     }
 }