+++ OUDS/Core/Components/Sources/Tags/Internal/TagLabelStyle+TagShapeModifier.swift
@@ -2,13 +2,13 @@
 // Software Name: OUDS iOS
 // SPDX-FileCopyrightText: Copyright (c) Orange SA
 // SPDX-License-Identifier: MIT
-// 
+//
 // This software is distributed under the MIT license,
 // the text of which is available at https://opensource.org/license/MIT/
 // or see the "LICENSE" file for more details.
-// 
+//
 // Authors: See CONTRIBUTORS.txt
-// Software description: A SwiftUI components library with code examples for Orange Unified Design System 
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
 //
 
 import OUDSTokensSemantic
@@ -16,9 +16,10 @@ import SwiftUI
 
 struct TagShapeModifier: ViewModifier {
 
-    // MARK: Stored Properties
+    // MARK: Stored properties
 
     let shape: OUDSTag.Shape
+
     @Environment(\.theme) private var theme
 
     // MARK: Body
@@ -33,9 +34,9 @@ struct TagShapeModifier: ViewModifier {
     private var cornerRadius: CGFloat {
         switch shape {
         case .square:
-            return theme.borders.borderRadiusNone
+            theme.borders.borderRadiusNone
         case .rounded:
-            return theme.tag.tagBorderRadius
+            theme.tag.tagBorderRadius
         }
     }
 }