+++ OUDS/Core/Components/Sources/Tags/OUDSTag.swift
@@ -15,15 +15,12 @@ import OUDSFoundations
 import OUDSTokensSemantic
 import SwiftUI
 
-// MARK: - OUDS Tag
-
-/// A tag is a small element that shows short info like a label, keyword, or category.
+/// A tag is a small element that shows short information like a label, keyword, or category.
 /// It helps users quickly find, group, or understand content.
 ///
-/// Tags have seven status depending on the context of the information they represent. Each state is designed
-/// to convey a specific meaning and ensure clarity in communication.
-///
 /// ## Hierarchies
+///
+/// Tags have two hierarchies so as to highlight or not their content.
 ///
 /// - **emphasized**: A tag with a solid, high-contrast background. Used to draw strong attention to important
 /// labels or categories. Emphasized tags stand out prominently against the interface and are ideal for
@@ -35,57 +32,63 @@ import SwiftUI
 ///
 /// ## Status
 ///
-/// Badges have seven status depending on the context of the information they represent.
-/// Each state is designed to convey a specific meaning and ensure clarity in communication.
+/// Tags have seven status depending on the context of the information they represent. Each state is designed
+/// to convey a specific meaning and ensure clarity in communication.
 ///
-/// - **Neutral** Default or inactive state. Used for standard labels, categories, or when no specific
+/// - **Neutral**: Default or inactive state. Used for standard labels, categories, or when no specific
 /// status needs to be communicated.
 ///
-/// - **Accent** Used to draw attention to new features, recommendations, or content suggestions.
+/// - **Accent**: Used to draw attention to new features, recommendations, or content suggestions.
 /// Invites users to explore and engage with new offerings, creating an exciting and engaging experience.
 ///
-/// - **Positive** Indicates success, confirmation, or a positive status. Commonly used to highlight
+/// - **Positive**: Indicates success, confirmation, or a positive status. Commonly used to highlight
 /// completed actions or approved items.
 ///
-/// - **Warning** Signals caution or a potentially risky situation. Used to draw attention to items
+/// - **Warning**: Signals caution or a potentially risky situation. Used to draw attention to items
 /// requiring user awareness or intervention.
 ///
-/// - **Negative** Represents errors, critical issues, or urgent attention needed. Used to highlight
+/// - **Negative**: Represents errors, critical issues, or urgent attention needed. Used to highlight
 ///  problems or failed actions.
 ///
-/// - **Info** Conveys informational messages or supplementary details. Used for neutral, helpful,
+/// - **Info**: Conveys informational messages or supplementary details. Used for neutral, helpful,
 /// or contextual information.
 ///
-/// - **Disabled** Shows that the tag is inactive and cannot be interacted with. Appears faded or greyed out. This is not
-///  allowed when loader activated.
+/// - **Disabled**: Shows that the tag is inactive and cannot be interacted with. Appears faded or greyed out. This is not
+///  allowed when loader is activated.
 ///
 /// ## Shape
 ///
-/// - **Rounded** A tag with fully rounded corners, creating a pill-shaped appearance. Rounded
+/// Tags can have two shapes.
+///
+/// - **Rounded**: A tag with fully rounded corners, creating a pill-shaped appearance. Rounded
 /// tags offer a softer and more approachable look, suitable for most modern interfaces.
 ///
-/// - **Square** A tag with sharp, square corners. Squared tags provide a more formal, structured, or
+/// - **Square**: A tag with sharp, square corners. Squared tags provide a more formal, structured, or
 /// technical feel. They are often used in business contexts to label promotions, offers, or important notices.
 ///
 /// ## Size
 ///
-/// - **Default** The standard tag size, suitable for most use cases and offering good readability.
+/// Tags can have two sizes.
 ///
-/// - **Small** A compact tag with reduced height and font size. Used when saving space is important or
+/// - **Default**: The standard tag size, suitable for most use cases and offering good readability.
+///
+/// - **Small**: A compact tag with reduced height and font size. Used when saving space is important or
 /// when grouping elements visually.
 ///
 /// ## Layout
 ///
-/// -**Text only** A tag that displays only text. Used for simple labels, categories, or keywords without additional visual
+/// There are four available layouts for tags.
+///
+/// -**Text only**: A tag that displays only text. Used for simple labels, categories, or keywords without additional visual
 ///  elements.
 ///
-///  -**Text + Bullet** A tag with a small indicator (dot) alongside the text. Used to show status, presence, or activity
+///  -**Text + Bullet**: A tag with a small indicator (dot) alongside the text. Used to show status, presence, or activity
 ///  next to the label.
 ///
-/// -**Text + Icon** A tag that includes an icon before the text. Used to visually reinforce the meaning of the tag,
+/// -**Text + Icon**: A tag that includes an icon before the text. Used to visually reinforce the meaning of the tag,
 /// such as status, type, or action.
 ///
-/// - **Text + Loader** A tag that combines a loading spinner (or progress indicator) with text. Used to indicate that
+/// - **Text + Loader**: A tag that combines a loading spinner (or progress indicator) with text. Used to indicate that
 /// a process or action related to the tag is in progress.
 ///
 /// ## Code samples
@@ -93,6 +96,8 @@ import SwiftUI
 /// ```swift
 ///     // Text only with neutral status, for emphasized hierarchy with rounded shape in default size
 ///     OUDSTag(label: "Label", hierarchy: .emphasized, status: .neutral, shape: .rounded, size: .default)
+///     // Or also
+///     OUDSTag(label: "Label")
 ///
 ///     // Text with bullet and negative status, using default hierarchy (emphasized), shape (rounded) and size (default)
 ///     OUDSTag(label: "Label", hasBullet: true, status: .negative)
@@ -105,13 +110,12 @@ import SwiftUI
 /// ```
 ///
 /// ## Design documentation
-///     @TODO: Update doc url)
 ///
 /// [unified-design-system.orange.com](https://unified-design-system.orange.com/)
 ///
 /// - Version: 1.1.0 (Figma component design version)
 /// - Since: 0.18.0
-public struct OUDSTag: View {
+public struct OUDSTag: View { // TODO: #408 - Add documentation hyperlink in doc above
 
     // MARK: Stored Properties
 
@@ -123,28 +127,30 @@ public struct OUDSTag: View {
 
     @Environment(\.theme) private var theme
 
+    // MARK: - Configuration enums
+
     enum `Type` {
         case textAndIcon(label: String, icon: OUDSTag.Icon?)
         case textAndLoader(label: String)
 
         var label: String {
             switch self {
-            case .textAndIcon(let label, _), .textAndLoader(let label):
-                return label
+            case let .textAndIcon(label, _), let .textAndLoader(label):
+                label
             }
         }
     }
 
-    /// Type of iccon in a Tag.
+    /// Type of icon in an `OUDSTag`
     public enum Icon {
-        /// The bullet icon in `OUDSTag`
+        /// The tag has a bullet as icon
         case bullet
 
-        /// An icon from asset
+        /// The tag has an image as icon
         case asset(Image)
     }
 
-    /// Represents the hierarchy of a tag
+    /// Represents the hierarchy of an `OUDSTag`
     public enum Hierarchy {
         /// A tag with a solid, high-contrast background.
         /// Used to draw strong attention to important labels or categories. Emphasized tags stand out
@@ -157,7 +163,7 @@ public struct OUDSTag: View {
         case muted
     }
 
-    /// Defines the shape of the tag
+    /// Defines the shape of an `OUDSTag`
     public enum Shape {
         /// A tag with sharp, square corners.
         /// Squared tags provide a more formal, structured, or technical feel. They are often used in business contexts to label promotions, offers, or important notices.
@@ -168,6 +174,7 @@ public struct OUDSTag: View {
         case rounded
     }
 
+    /// Defines the size of an `OUDSTag`
     public enum Size {
         /// The standard tag size, suitable for most use cases and offering good readability.
         case `default`
@@ -201,41 +208,28 @@ public struct OUDSTag: View {
         case disabled
     }
 
-    // MARK: Initializers
+    // MARK: - Initializer
 
-    /// Create a tag with simple label, categorie, or keyword without additional visual elements.
-    /// A small indicator (bullet) could be displays to show status, presence, or activity next to the label.
-    ///
-    ///  Four different layouts are supported:
-    ///     - **Text only**: when `icon` is `nil`, the tag displays only text.
-    ///       Used for simple labels, categories, or keywords without additional visual elements.
-    ///
-    ///     - **Text and bullet**: when `icon` is equal to `OUDSTag.Icon.bullet`, the tag displays a small
-    ///     indicator (bullet) alongside the text. Used to show status, presence, or activity next to the label.
-    ///
-    ///     - **Text and icon**: when `icon` is not `nil`, the tag includes an icon from asset before the text.
-    ///     Used to visually reinforce the meaning of the tag, such as status, type, or action.
-    ///
-    ///     - **Text and loader**: when `loader` is `true`, the tag combines a loading spinner (or progress indicator)
-    ///     with text. Used to indicate that a process or action related to the tag is in progress.
+    /// Create a tag with simple label and mayve an icon.
     ///
     /// - Parameters:
     ///    - label: The label displayed in the tag
     ///    - icon: The icon displayed before the label, or `nil` if there is no icon
-    ///    - hierarchy: The importance of the tag. Its background color and its content color are based on this hierarchy combined to the [status] of the tag
+    ///    - hierarchy: The importance of the tag. Its background color and its content color are based on this hierarchy combined to the `status` of the tag
     ///    - status: The status of the tag. Its background color and its content color are based on this status combined to
     ///    the `hierarchy` of the tag
     ///    - shape: The shape of the tag that allows to play with its corners appearance.
     ///    - size: The size of the tag
-    ///    - loader: An optional loader (or progress indicator) displayed before the **label**. It will
-    ///    replace the `icon` if provided. Used to indicate that a process or action related to the tag is in progress.
+    ///    - loader: An optional loader (or progress indicator) displayed before the `label`. It will
+    ///    replace the `icon` if provided.
     public init(label: String,
                 icon: OUDSTag.Icon? = nil,
                 hierarchy: Hierarchy = .emphasized,
                 status: Status = .neutral,
                 shape: Shape = .rounded,
                 size: Size = .default,
-                loader: Bool = false) {
+                loader: Bool = false)
+    {
         self.hierarchy = hierarchy
         self.status = status
         self.shape = shape
@@ -244,9 +238,9 @@ public struct OUDSTag: View {
             if status == .disabled {
                 OL.fatal("An OUDSTag with OUDSTag.Status.disabled status has been detected with loader activated, which is not allowed.")
             }
-            self.type = .textAndLoader(label: label)
+            type = .textAndLoader(label: label)
         } else {
-            self.type = .textAndIcon(label: label, icon: icon)
+            type = .textAndIcon(label: label, icon: icon)
         }
     }
 
@@ -263,10 +257,10 @@ public struct OUDSTag: View {
 
     private var hasIcon: Bool {
         switch type {
-        case .textAndIcon(_, let icon):
-            return icon != nil
+        case let .textAndIcon(_, icon):
+            icon != nil
         default:
-            return true
+            true
         }
     }
 }